backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Real-time Communication
python-socketio==5.11.0
python-engineio==4.9.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Payment Integration
stripe==7.11.0
razorpay==1.4.1

# AI & ML Dependencies
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Caching
redis==5.0.1

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.6

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
